Moved mocks directory
