+++ src/__mocks__/@mapbox/mapbox-sdk/services/geocoding.ts
