+++ src/__mocks__/aws-sdk/clients/dynamodb.ts
